feat: Adds scheduled task for refreshing features
Currently not validating the URL, and silently failing when we fail to
parse the response from the fetched data. But the tokio tasks are in
process.

Should probably add a last_fetch for each token, so that we can
separately configure which ones to fetch, and then reduce the sleep for
the token refresher.